backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
anthropic==0.40.0
chromadb==0.5.23
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.7.0
python-multipart==0.0.20
pymupdf==1.24.14
pdfplumber==0.11.4
Pillow==11.0.0
python-dotenv==1.0.1
httpx==0.28.1
aiofiles==24.1.0
numpy==2.2.0
PyJWT==2.11.0
reportlab==4.4.10
bcrypt==5.0.0
apscheduler==3.11.2
cachetools==7.0.1
sendgrid==6.12.5
stripe==14.4.0
